learning_rest1/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
country-converter==1.2
countryflag==0.1.2b3
defusedxml==0.7.1
-e git+ssh://git@github.com/django/django.git@89e28e13ecbf9fbcf235e16d453c08bbf2271244#egg=Django
django-filter==24.3
djangorestframework==3.15.2
emoji==2.13.2
emoji-country-flag==2.0.1
httpie==3.2.4
idna==3.10
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.1
pandas==2.2.3
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
